+++ packages/core/src/components/DateRangePicker/DateRangeTextFields/DateRangeTextFields.tsx
@@ -1,9 +1,10 @@
-import { DateRangeIcon } from '@medly-components/icons';
+import { ChevronDownIcon, DateRangeIcon } from '@medly-components/icons';
 import React from 'react';
-import { DateIconWrapper } from '../../DatePicker/DatePicker.styled';
+import { DateIconWrapper as IconWrapper } from '../../DatePicker/DatePicker.styled';
 import datePickerPattern from '../../DatePicker/datePickerPattern';
 import { HelperAndErrorTextTooltip } from '../../HelperAndErrorTextTooltip/HelperAndErrorTextTooltip';
 import * as TextFieldStyled from '../../TextField/Styled';
+import { PopoverTypes } from '../types';
 import DateRangeTextField from './DateRangeTextField';
 import { Wrapper } from './DateRangeTextFields.styled';
 import InputSeparator from './InputSeparator';
@@ -26,7 +27,10 @@ export const DateRangeTextFields: React.FC<Props> = React.memo(props => {
             displayFormat,
             startDateLabel,
             endDateLabel,
-            showTooltipForHelperAndErrorText
+            showTooltipForHelperAndErrorText,
+            showChevronIcon,
+            activePopover,
+            onCustomRangeIconClick
         } = props,
         {
             mask,
@@ -66,9 +70,15 @@ export const DateRangeTextFields: React.FC<Props> = React.memo(props => {
         };
 
     const Prefix = () => (
-        <DateIconWrapper {...iconProps}>
+        <IconWrapper {...iconProps}>
             <DateRangeIcon id={`${id}-calendar-icon`} onClick={onIconClick} size={size} />
-        </DateIconWrapper>
+        </IconWrapper>
+    );
+
+    const Suffix = () => (
+        <IconWrapper {...iconProps}>
+            <ChevronDownIcon id={`${id}-custom-date-range-options-icon`} onClick={onCustomRangeIconClick} size={size} />
+        </IconWrapper>
     );
 
     return (
@@ -81,6 +91,8 @@ export const DateRangeTextFields: React.FC<Props> = React.memo(props => {
                 isErrorPresent={isErrorPresent}
                 isLabelPresent
                 isActive={isActive}
+                showChevronIcon={showChevronIcon}
+                areCustomOptionsVisible={isActive && activePopover === PopoverTypes.CUSTOM_RANGE_OPTIONS}
                 onBlur={validateOnWrapperBlur}
             >
                 {showDecorators && (
@@ -110,6 +122,11 @@ export const DateRangeTextFields: React.FC<Props> = React.memo(props => {
                 />
                 {showTooltipForHelperAndErrorText && (
                     <HelperAndErrorTextTooltip id={id} errorText={errorText || builtInErrorMessage} helperText={helperText} />
+                )}
+                {showChevronIcon && (
+                    <TextFieldStyled.Suffix size={size}>
+                        <Suffix />
+                    </TextFieldStyled.Suffix>
                 )}
             </Wrapper>
             {showDecorators && !showTooltipForHelperAndErrorText && (